splash screen + vulkan validation fixes

+++ Lumos/Source/Lumos/Audio/Sound.cpp
@@ -45,7 +45,7 @@ namespace Lumos
 
         uint32_t j = 0;
         uint32_t channelSize = channels * bytesPerSample;
-        for(uint32_t i = 0; i < dataSize; i += channelSize)
+        for(uint32_t i = 0; i < uint32_t(dataSize); i += channelSize)
         {
             switch(bytesPerSample)
             {
